feat: improve btc things
