+++ packages/swapper/src/swappers/ChainflipSwapper/utils/getUtxoTxFees.ts
@@ -2,6 +2,8 @@ import type { GetFeeDataInput, UtxoChainAdapter } from '@shapeshiftoss/chain-ada
 import type { UtxoChainId } from '@shapeshiftoss/types'
 import { bn } from '@shapeshiftoss/utils'
 
+import type { QuoteFeeData } from '../../../types'
+
 type GetUtxoTxFeesInput = {
   sellAmountCryptoBaseUnit: string
   sellAdapter: UtxoChainAdapter
@@ -12,7 +14,7 @@ export const getUtxoTxFees = async ({
   sellAmountCryptoBaseUnit,
   sellAdapter,
   publicKey,
-}: GetUtxoTxFeesInput): Promise<string> => {
+}: GetUtxoTxFeesInput): Promise<Omit<QuoteFeeData, 'protocolFees'>> => {
   const getFeeDataInput: GetFeeDataInput<UtxoChainId> = {
     // One of many vault addresses - just used as a placeholder for the sake of loosely estimating fees - we *need* a *to* address for simulation or this will throw
     to: 'bc1pfh5x55a3v92klcrdy5yv6yrt7fzr0g929klkdtapp3njfyu4qsyq8qacyf',
@@ -24,5 +26,11 @@ export const getUtxoTxFees = async ({
 
   const feeData = feeDataOptions['fast']
 
-  return bn(feeData.txFee).dp(0).toString()
+  return {
+    networkFeeCryptoBaseUnit: feeData.txFee,
+    chainSpecific: {
+      satsPerByte: feeData.chainSpecific.satoshiPerByte,
+      byteCount: bn(feeData.txFee).dividedBy(feeData.chainSpecific.satoshiPerByte).dp(0).toString(),
+    },
+  }
 }
